feat: moved new views to separate files
