+++ VultisigApp/VultisigApp/Views/Send/SendCryptoDone/SendCryptoDoneContent.swift
@@ -1,0 +1,18 @@
+//
+//  SendCryptoDoneContent.swift
+//  VultisigApp
+//
+//  Created by Gaston Mazzeo on 22/07/2025.
+//
+
+struct SendCryptoContent {
+    let coin: Coin
+    let amountCrypto: String
+    let amountFiat: String
+    let hash: String
+    let explorerLink: String
+    
+    let fromAddress: String
+    let toAddress: String
+    let fee: (crypto: String, fiat: String)
+}
